moved feauterd movie file to components folder and deleted unnecessary import
